Update README.md
This repository hosts the source code for an event management system website built using the MVC design pattern. The project incorporates the Unit of Work pattern and is organized using areas for modularity. The system supports user registration, role-based authorization for admins and organizers, and integrates Stripe for secure payment processing. The website is fully responsive, ensuring a seamless experience across devices.